feat: TabBarModifier 적용
