+++ Projects/Feature/ProfileSetup/Interface/Sources/IntroductionSetup/IntroductionSetupView.swift
@@ -40,6 +40,7 @@ public struct IntroductionSetupView: View {
         }
       }
     }
+    .toolbar(.hidden, for: .tabBar)
     .onLoad {
       store.send(.onLoad)
       isVisibleTabBar.toggle()
