+++ Projects/Feature/TabBar/Interface/Sources/TabBarModifier.swift
@@ -8,37 +8,7 @@
 import Foundation
 import SwiftUI
 
-public enum TabType: Hashable, CaseIterable {
-  case sandBeach
-  case bottleStorage
-  case myPage
-  
-  var title: String {
-    switch self {
-    case .sandBeach:
-      return "모래사장"
-      
-    case .bottleStorage:
-      return "보틀 보관함"
-      
-    case .myPage:
-      return "마이페이지"
-    }
-  }
-  
-  var image: Image.BottleImageSystem {
-    switch self {
-    case .sandBeach:
-      return .icom(.sandBeach)
-      
-    case .bottleStorage:
-      return .icom(.bottleStorage)
-      
-    case .myPage:
-      return .icom(.myPage)
-    }
-  }
-}
+import SharedDesignSystem
 
 private struct TabBarModifier: ViewModifier {
   
@@ -81,6 +51,7 @@ private struct TabBarModifier: ViewModifier {
         }
         // TODO: - DesignSystem -> 디자인들 Contants 관리
         .frame(height: 106)
+        .border(.blue)
         .background {
           RoundedRectangle(cornerRadius: BottleRadiusType.xl.value)
             .fill(ColorToken.background(.secondary).color)
